requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
PyPDF2
python-dotenv
google-genai
